requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
groq
plotly>=5.18.0
python-dotenv>=1.0.0
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-search-results==2.4.2
serpapi